Change to double precision

+++ source/headers/rgb.cpp
@@ -2,9 +2,9 @@
 
 RGB::RGB(void)
     : r(0), g(0), b(0) {}
-RGB::RGB(float c)
+RGB::RGB(double c)
     : r(c), g(c), b(c) {}
-RGB::RGB(float _r, float _g, float _b)
+RGB::RGB(double _r, double _g, double _b)
     : r(_r), g(_g), b(_b) {}
 
 bool RGB::operator != (const RGB &c)
